+++ lib/tcp_server/utils/static_file_handler.dart
@@ -1,9 +1,9 @@
 part of tcp_server;
 
 class StaticFileHandler {
-  final String basePath;
+  final List filePaths;
 
-  StaticFileHandler(this.basePath);
+  StaticFileHandler(this.filePaths);
 
   _send404(HttpResponse response) {
     response.statusCode = HttpStatus.NOT_FOUND;
@@ -11,20 +11,36 @@ class StaticFileHandler {
     response.close();
   }
 
-  // TODO: etags, last-modified-since support
   onRequest(HttpRequest request) {
-    //print("incoming file request");
     final String localPath =
         request.uri.path == '/' ? 'asteroidracers.html' : request.uri.path.substring(1, request.uri.path.length);
+
+    bool fileFound = false;
+    for(String basePath in filePaths){
+      String filePath = path.join(basePath, localPath);
+      print("file request for path $filePath");
+      File file = new File(filePath);
+      if(file.existsSync()){
+        sendFileResponse(file, request);
+        fileFound = true;
+        break;
+      }
+    }
     
-    String filePath = path.join(basePath, localPath);
-    final File file = new File(filePath);
-    if(file.existsSync()){
-        file.openRead().pipe(request.response).catchError((e) => print(e));
-    }
-    else {
+    if(!fileFound){
       _send404(request.response);
-      print("failed to load file $filePath");
+      print("failed to load file $localPath");
     }
+  }
+  
+  void sendFileResponse(File file, HttpRequest request){
+    // MIME type
+    String mimeType = mime(file.path);
+    if (mimeType == null) mimeType = 'text/plain; charset=UTF-8';
+    request.response.headers.set('Content-Type', mimeType);
+    // Content Length
+    request.response.contentLength = file.lengthSync();
+    // send response
+    file.openRead().pipe(request.response).catchError((e) => print(e));
   }
 }