+++ lib/web_client/views/ui/button.dart
@@ -12,6 +12,8 @@ class Button extends stagexl.Sprite {
   
   Button([num width = 100, num height = 100]){
     
+    this.width = width;
+    
     stagexl.Sprite redRect = _rect(width, height, stagexl.Color.Red);
     stagexl.Sprite darkBlueRect = _rect(width, height, stagexl.Color.DarkBlue);
     stagexl.Sprite darkRedRect = _rect(width, height, stagexl.Color.DarkRed);
@@ -21,14 +23,21 @@ class Button extends stagexl.Sprite {
       ..width = width
       ..defaultTextFormat.align = stagexl.TextFormatAlign.CENTER;
     _textField.y = height/2 -_textField.textHeight/2;
+    _textField.height = _textField.textHeight;
     
     stagexl.SimpleButton button = new stagexl.SimpleButton(darkRedRect);   
     button.downState = darkBlueRect;
     button.overState = redRect;
     button.hitTestState = redRect;
-    
+        
+    print(this.height);
+
     this.addChild(button);
+    print(this.height);
+
     this.addChild(_textField);
+    
+    print(this.height);
   }
   
   _rect(num width, num height, int color){