+++ lib/web_client/views/ui/ui_helper.dart
@@ -2,23 +2,24 @@ part of web_client;
 
 class UIHelper {
   
-  static num _textFieldHeight = 20;
   
   static stagexl.TextField createTextField({String text: "", num numLines: 1}){
     stagexl.TextField textField = new stagexl.TextField();
     textField.textColor = stagexl.Color.White;
-    textField.height = _textFieldHeight * numLines;
+    textField.height = textField.textHeight * numLines;
     textField.text = text;
+    
     return textField;
   }
 
   static stagexl.TextField createInputField([num numLines = 1]){
     stagexl.TextField textField = new stagexl.TextField();
     textField.backgroundColor = stagexl.Color.White;
-    textField.textColor = stagexl.Color.Black;
     textField.background = true;  
-    textField.height = _textFieldHeight * numLines;
+    textField.textColor = stagexl.Color.Black;
+    textField.height = textField.textHeight * numLines + 5;
     textField.type = stagexl.TextFieldType.INPUT;
+    
     return textField;
   }
 }